mcp: carve tool_params.rs (pure move)
The tool parameter and response-schema types (every *Params struct,
McpResponseSchema and friends) plus the persisted-log readers behind
get_logs: 787 body lines and 1 test verbatim. The tool tests that
merely construct params stay with the IntendantServer methods they
exercise. Usual glue; purity-checked; mcp:: tests 178/178.
